// src/models/spotify/spotify-episode.d.ts
// Generated by https://quicktype.io

import { Show } from "./spotify-show";

export interface Episode {
  audio_preview_url:      string;
  description:            string;
  duration_ms:            number;
  explicit:               boolean;
  external_urls:          ExternalUrls;
  href:                   string;
  id:                     string;
  images:                 Image[];
  is_externally_hosted:   boolean;
  is_playable:            boolean;
  language:               string;
  languages:              string[];
  name:                   string;
  release_date:           string;
  release_date_precision: string;
  show?:                  Show;
  type:                   string;
  uri:                    string;
}

export interface ExternalUrls {
  spotify: string;
}

export interface Image {
  height: number;
  url:    string;
  width:  number;
}