// src/models/spotify/spotify-show.d.ts
// Generated by https://quicktype.io

import { Episode } from "./spotify-episode";

export interface ShowPaginator {
  href:     string;
  items:    Episode[];
  limit:    number;
  next:     string;
  offset:   number;
  previous: null;
  total:    number;
}

export interface Show {
  available_markets:    string[];
  copyrights:           any[];
  description:          string;
  episodes:             ShowPaginator;
  explicit:             boolean;
  external_urls:        ExternalUrls;
  href:                 string;
  id:                   string;
  images:               Image[];
  is_externally_hosted: boolean;
  languages:            string[];
  media_type:           string;
  name:                 string;
  publisher:            string;
  total_episodes:       number;
  type:                 string;
  uri:                  string;
}

export interface ExternalUrls {
  spotify: string;
}

export interface Image {
  height: number;
  url:    string;
  width:  number;
}